requirements.txt
# EnsinaLab Content Engine - Requirements

# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Celery (async tasks)
celery==5.3.4
redis==5.0.1

# OpenAI
openai==1.3.7

# LangChain (RAG)
langchain==0.0.340
langchain-openai==0.0.2

# Vector DB (escolher uma)
# qdrant-client==1.7.0
# pinecone-client==2.2.4

# Video generation
moviepy==1.0.3
Pillow==10.1.0
numpy==1.26.2

# TTS (escolher uma opção)
# elevenlabs==0.2.27  # ElevenLabs TTS
# boto3==1.29.7  # Amazon Polly
# TTS==0.20.0  # Coqui TTS (open-source)

# Utils
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.2
